[Refactor]combined login api
