+++ src/pages/api/rest.ts
@@ -1,11 +1,26 @@
 import axios from "axios";
 import Cookies from "js-cookie";
+import {instance} from "@/pages/api/instance";
 
-export const instance = axios.create({
-    baseURL: process.env.NEXT_PUBLIC_API_URL,
-});
+// export const instance = axios.create({
+//     baseURL: process.env.NEXT_PUBLIC_API_URL,
+// });
 
 // KakaoLogin
+// export const getKakaoLogin = async (KAKAO_CODE: string) => {
+//     try {
+//         const response = await instance.post(`/api/kakao?code=${KAKAO_CODE}`);
+//         const accessToken = response.headers.authorization;
+//         const refreshToken = response.headers.refreshtoken;
+//         Cookies.set("access_Token", accessToken, { path: "/" });
+//         Cookies.set("refresh_Token", refreshToken, { path: "/" });
+//         console.log("카카오 로그인 성공", response);
+//         return response.data;
+//     } catch (error) { // 로그인 에러 처리
+//         throw error;
+//     }
+// };
+
 export const getKakaoLogin = async (KAKAO_CODE: string) => {
     try {
         const response = await instance.post(`/api/kakao?code=${KAKAO_CODE}`);
@@ -19,3 +34,21 @@ export const getKakaoLogin = async (KAKAO_CODE: string) => {
         throw error;
     }
 };
+
+// test용 로그인
+export const credentialLogin = async () => {
+    try {
+        const response = await axios.post(`${process.env.NEXT_PUBLIC_API_URL}/auth/login`, {
+            nickname: "user", password: "user" ,
+        });
+        const accessToken = response.headers.authorization;
+        const refreshToken = response.headers.refreshtoken;
+        Cookies.set("access_Token", accessToken, { path: "/" });
+        Cookies.set("refresh_Token", refreshToken, { path: "/" });
+        console.log(response)
+        return response
+    } catch (error) {
+        console.error('Login error:', error);
+        throw error;
+    }
+}
